* commit 'd39268dd9512ce5d31dede61ddeea43a6d33231b':
  Working with statements

+++ src/main/java/com/endava/service_system/model/dto/PaymentDto.java
@@ -2,9 +2,11 @@ package com.endava.service_system.model.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PaymentDto {
     private Long C;
-    private double S;
+    private BigDecimal S;
     private String D;
 }
